pyproject.toml
[project]
name = "aicontent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.75",
    "langchain-openai>=0.3.32",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "serpapi>=0.1.5",
]
